requirements.txt
# Function dependencies, for example:
# package>=version
# langchain
gpt_index
line-bot-sdk
requests